Validate org-user membership from gateway
Adding credential and execution environment roles
validates that the user belongs to the same org
as the credential or EE.

In some situations, the user-org membership has not
yet been synced from gateway to controller.

In this case, controller will make a request to
gateway to check if the user is part of the org.

Signed-off-by: Seth Foster <fosterbseth@gmail.com>

rename gateway to resource_server

Signed-off-by: Seth Foster <fosterbseth@gmail.com>

use dab fork

Signed-off-by: Seth Foster <fosterbseth@gmail.com>